ml/requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2
SQLAlchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
